Calculate supervised CCA on three data matrices

getCCA3 starts the supervised sparse CCA with elastic net for 2 biological data sets and 1 design data set Z. Data sets X and Y are strongly regularized (ridge regression); Z if neccessary (det(Z)==0).

getCCA3(X,Y,Z, end=c(0.3,0.5,3), step=c(0.01,0.01,0.2), numCV=10,
n.r=10, max.counter.test=10)

Arguments

X

biological data set with dim(X): n rows, p features

Y

biological data set with dim(Y): n rows, q features

Z

design data set with dim(Z): n rows, r features; describes experimental design with binary vectors

end

vector containing the three maximal lambda-values for the grid search

step

vector containing the step size for grid search for the lambdas

numCV

number of canonical variables, which will be calculated

n.r

number of resampling runs, 10 is okay

max.counter.test

number of random start vectors

Details

Value

cc3.weight.x, cc3.weight.y, cc3.weight.z

Matrix with weights of canonical variables for data set X, Y and Z respectively (columnwise)

ccc3.CV.x, ccc3.CV.y, ccc3.CV.z

Matrix with canonical variables for data set X, Y and Z respectively (columnwise)

corr

Vector with absolute correlation coefficients of canonical variables (columnwise): (corr(cv(X),cv(Z))+ corr(cv(Y),cv(Z))/2

lambda

Matrix with 3 parameters lambda for each data set and each canonical variable (columns)

numCV

Number of canonical variables

Author

Andrea Thum

Note

Examples